requirements.txt
alembic==1.16.5
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
Deprecated==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
future @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
macholib @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysql-connector-python==9.4.0
numpy==2.0.2
ordered-set==4.1.0
packaging==24.2
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
qrcode==8.2
razorpay==2.0.0
requests==2.32.5
rich==13.9.4
six @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==2.0.48
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
zipp==3.23.0
